extract common adapters module and create common http configuration
