+++ knotx-adapters/knotx-adapter-action-http/pom.xml
@@ -44,6 +44,17 @@
     </dependency>
 
     <dependency>
+      <groupId>com.cognifide.knotx</groupId>
+      <artifactId>knotx-adapter-api</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.cognifide.knotx</groupId>
+      <artifactId>knotx-adapter-common</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
       <groupId>ch.qos.logback</groupId>
       <artifactId>logback-classic</artifactId>
       <version>1.1.3</version>
