+++ knotx-adapters/knotx-adapter-common/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Knot.x - Reactive microservice assembler - HTTP Service Adapter
+    Knot.x - Reactive microservice assembler - Adapter Common
 
     Copyright (C) 2016 Cognifide Limited
 
@@ -28,14 +28,8 @@
     <version>0.6.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>knotx-adapter-service-http</artifactId>
-  <name>Knot.x - Reactive microservice assembler - HTTP Service Adapter</name>
-
-
-  <properties>
-    <!-- the main verticle class name -->
-    <main.verticle>com.cognifide.knotx.adapter.service.http.HttpServiceAdapterVerticle</main.verticle>
-  </properties>
+  <artifactId>knotx-adapter-common</artifactId>
+  <name>Knot.x - Reactive microservice assembler - Adapter Common</name>
 
   <dependencies>
     <dependency>
@@ -46,11 +40,6 @@
     <dependency>
       <groupId>com.cognifide.knotx</groupId>
       <artifactId>knotx-adapter-api</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.cognifide.knotx</groupId>
-      <artifactId>knotx-adapter-common</artifactId>
       <version>${project.version}</version>
     </dependency>
 
@@ -70,34 +59,14 @@
       <artifactId>vertx-rx-java</artifactId>
     </dependency>
     <dependency>
-      <groupId>io.vertx</groupId>
-      <artifactId>vertx-hazelcast</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.reactivex</groupId>
-      <artifactId>rxjava</artifactId>
+      <groupId>com.google.guava</groupId>
+      <artifactId>guava</artifactId>
     </dependency>
 
-    <!-- REPOSITORY -->
-    <dependency>
-      <groupId>org.jsoup</groupId>
-      <artifactId>jsoup</artifactId>
-    </dependency>
-
-    <!-- HANDLEBARS -->
+    <!-- Handlebars -->
     <dependency>
       <groupId>com.github.jknack</groupId>
       <artifactId>handlebars</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.github.jknack</groupId>
-      <artifactId>handlebars-jackson2</artifactId>
-    </dependency>
-
-    <!-- COMMON -->
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
     </dependency>
 
     <!-- Tests -->
@@ -106,39 +75,8 @@
       <artifactId>junit</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.powermock</groupId>
-      <artifactId>powermock-api-mockito</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.hamcrest</groupId>
-      <artifactId>hamcrest-all</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.powermock</groupId>
-      <artifactId>powermock-module-junit4</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-all</artifactId>
-    </dependency>
-    <dependency>
       <groupId>io.vertx</groupId>
       <artifactId>vertx-unit</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.cognifide.knotx</groupId>
-      <artifactId>knotx-common</artifactId>
-      <version>${project.version}</version>
-      <type>test-jar</type>
-      <scope>test</scope>
-    </dependency>
-
-    <!-- TODO remove this dependency - move tests that require it to separate module with integration tests -->
-    <dependency>
-      <groupId>com.cognifide.knotx</groupId>
-      <artifactId>knotx-mocks</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
     </dependency>
   </dependencies>
 
@@ -154,14 +92,6 @@
             </goals>
           </execution>
         </executions>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-shade-plugin</artifactId>
-      </plugin>
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>exec-maven-plugin</artifactId>
       </plugin>
     </plugins>
   </build>
