+++ knotx-adapters/pom.xml
@@ -35,6 +35,7 @@
 
   <modules>
     <module>knotx-adapter-api</module>
+    <module>knotx-adapter-common</module>
     <module>knotx-adapter-service-http</module>
     <module>knotx-adapter-action-http</module>
   </modules>
